Initial Data Load

+++ cache.js
@@ -1,9 +1,10 @@
 const fs = require('fs');
 const path = require('path');
+const REQUEUED = 'requeued'
 
 function push (input, cacheDir, log) {
 	let data = JSON.stringify(input)
-	let fileName = Date.now()
+	let fileName = Date.now() + '.' + REQUEUED
 	fs.writeFileSync(`${cacheDir}/${fileName}.json`, data);
 	log(`Caching to ${fileName}`)
 }
@@ -13,14 +14,17 @@ function load (cacheDir, log) {
 	const fileArray = []
 	const directoryPath = cacheDir;
 	//passsing directoryPath and callback function
-	files = fs.readdirSync(directoryPath) 
+	files = fs.readdirSync(directoryPath)
+	for (i=0; i<files.length; i++)
+		if (files[i].includes(REQUEUED))
+			fileArray.push(files[i]) 
 	//console.log(files)
-	if (files.length == 0) {
+	if (fileArray.length == 0) {
 		log('No Files in Cache')
 		return false
 	}
 	else {
-		return files
+		return fileArray
 	}
 }
 
@@ -28,12 +32,14 @@ function send (filenames, cacheDir) {
 	var input = []
 	filenames.forEach(function (file) {
 		const thisfile = fs.readFileSync(`${cacheDir}/${file}`, 'utf8')
-  		input.push(...JSON.parse(thisfile));
+		if (thisfile.includes(REQUEUED))
+	  		input.push(...JSON.parse(thisfile));
 	});
 	
 	for (const file of filenames) {
-        fs.unlink((`${cacheDir}/${file}`), err => {
-            if (err) throw err;
+    	if (file.includes(REQUEUED))
+			fs.unlink((`${cacheDir}/${file}`), err => {
+				if (err) throw err;
         });
     };
